BWDO-524 update help screen
