+++ src/sc/branching_cli.py
@@ -25,40 +25,6 @@ logger = logging.getLogger(__name__)
 def cli():
     pass
 
-@cli.command()
-def init():
-    """Initialise for branching commands."""
-    SCBranching.init()
-
-@cli.command()
-def clean():
-    """Clean all modules. (git clean -fdx)."""
-    SCBranching.clean()
-
-@cli.command()
-def status():
-    """Show the working tree status."""
-    SCBranching.status()
-
-@cli.command()
-def reset():
-    """Clean and Reset all modules to remote manifest. (git reset --hard REMOTE)"""
-    SCBranching.reset()
-
-@cli.command()
-def build():
-    """Executes build related commands specified in a yaml file."""
-    pass
-
-@cli.command()
-@click.option("-n", "--no-checkout", is_flag=True, help="Do not checkout the branch.")
-@click.option("-b", "--force-broken", is_flag=True, help="Do not checkout the branch.")
-@click.option("-s", "--force-sync", is_flag=True, help="Do not checkout the branch.")
-def sync(no_checkout, force_broken, force_sync):
-    """Perform sync, and re-checkout branch"""
-    pass
-
-
 # Feature branch commands
 @cli.group()
 def feature():
@@ -293,82 +259,6 @@ def list():
 def checkout(name, force, verify):
     """Checkout a support branch."""
     SCBranching.checkout(BranchType.SUPPORT, name, force, verify)
-
-@cli.group()
-def tag():
-    pass
-
-@tag.command()
-def list():
-    """List tags in manifest or git repo."""
-    SCBranching.tag_list()
-
-@tag.command()
-@click.argument("tag")
-def show(tag):
-    """Show information about a tag in all repos."""
-    SCBranching.tag_show(tag)
-
-@tag.command()
-@click.argument("tag")
-def create(tag):
-    """Create a tag in all non READ_ONLY repos."""
-    SCBranching.tag_create(tag)
-
-@tag.command()
-@click.argument("tag")
-@click.option('-r', '--remote', is_flag=True, help="Remove in remotes as well as local.")
-def rm(tag, remote):
-    """Remove a tag in all non READ_ONLY repos."""
-    SCBranching.tag_rm(tag, remote)
-
-@tag.command()
-@click.argument("tag")
-def push(tag):
-    """Push given tag in all non READ_ONLY repos."""
-    SCBranching.tag_push(tag)
-
-@tag.command()
-@click.argument("tag")
-def check(tag):
-    """Check if a tag exists on all non READ_ONLY repos."""
-    SCBranching.tag_check(tag)
-
-@cli.group()
-def show():
-    """sc show commands."""
-    pass
-
-@show.command()
-def branch():
-    """Show the current status of branching."""
-    SCBranching.show_branch()
-
-@show.command()
-def repo_flow_config():
-    """Show git flow config for all projects."""
-    SCBranching.show_repo_flow_config()
-
-@show.command()
-def log():
-    SCBranching.show_log()
-
-@show.command()
-@click.argument("tag")
-def tag(tag):
-    """Show information about a tag."""
-    SCBranching.tag_show(tag)
-
-@show.command()
-def tags():
-    """List tags on the manifest."""
-    SCBranching.tag_list()
-
-@show.command()
-@click.argument("group", required=False)
-def group(group):
-    """List groups or show information about a group."""
-    SCBranching.group_show(group)
 
 if __name__ == '__main__':
     cli()
